A few small updates.

+++ Fused/MadgwickSensorFusion.m
@@ -212,7 +212,7 @@ static inline float invSqrt(float x)
         // Reference direction of Earth's magnetic field
         hx = mx * q0q0 - _2q0my * _q3 + _2q0mz * _q2 + mx * q1q1 + _2q1 * my * _q2 + _2q1 * mz * _q3 - mx * q2q2 - mx * q3q3;
         hy = _2q0mx * _q3 + my * q0q0 - _2q0mz * _q1 + _2q1mx * _q2 - my * q1q1 + my * q2q2 + _2q2 * mz * _q3 - my * q3q3;
-        _2bx = sqrt(hx * hx + hy * hy);
+        _2bx = sqrtf(hx * hx + hy * hy);
         _2bz = -_2q0mx * _q2 + _2q0my * _q1 + mz * q0q0 + _2q1mx * _q3 - mz * q1q1 + _2q2 * my * _q3 - mz * q2q2 + mz * q3q3;
         _4bx = 2.0f * _2bx;
         _4bz = 2.0f * _2bz;
